% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mingle.R
\name{buildQ.t.sub}
\alias{buildQ.t.sub}
\title{build the precision matrix for the distribution of \eqn{\boldsymbol{\mu}(t)|\boldsymbol{\mu}(t)\dots}. Intended mostly for internal use in simulation.}
\usage{
buildQ.t.sub(alpha, sigsq, c, wn.t)
}
\arguments{
\item{alpha}{scalar. alignment parameter from \eqn{[0, 1)}}

\item{sigsq}{scalar. marginal residual uncertainty after accounting for social effects.}

\item{c}{scalar. effective ego-network size for unconnected individuals.}

\item{wn.t}{matrix of dimension \code{c(n.indiv, n.indiv)} indicating network at time \eqn{t}}
}
\value{
matrix of dimension \code{c(n.indiv, n.indiv)}
}
\description{
build the precision matrix for the distribution of \eqn{\boldsymbol{\mu}(t)|\boldsymbol{\mu}(t)\dots}. Intended mostly for internal use in simulation.
}

