% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_mingle}
\alias{fit_mingle}
\title{fit the model using mcmc}
\usage{
fit_mingle(N_iterations, mu, phi, alpha, beta, p1, sigsq, c, w, alpha_p1,
  beta_p1, alpha_phi, beta_phi, alpha_alpha, beta_alpha, mu_beta, sigsq_beta,
  a_sigsq, b_sigsq, a_c, b_c, beta_p1_tune, beta_phi_tune, sigsq_alpha_tune)
}
\arguments{
\item{N_iterations}{if fitting all parameters, this takes approximately 1
minutes per 100 iterations on my laptop.}

\item{mu}{this is an array of dimension \code{c(n.indiv, T, 2, K)} where
\code{K} is some number of draws from the distribution of \eqn{s|\mu}.}

\item{phi}{vector of length \code{N_iterations}}

\item{alpha}{vector of length \code{N_iterations}}

\item{beta}{vector of length \code{N_iterations}}

\item{p1}{vector of length \code{N_iterations}}

\item{sigsq}{vector of length \code{N_iterations}}

\item{c}{vector of length \code{N_iterations}}

\item{w}{array of dimension \code{(n.indiv, n.indiv, T, N_iterations)}}

\item{alpha_p1}{hyperparameters}

\item{beta_p1}{hyperparameters}

\item{alpha_phi}{hyperparameters}

\item{beta_phi}{hyperparameters}

\item{alpha_alpha}{hyperparameters}

\item{beta_alpha}{hyperparameters}

\item{mu_beta}{hyperparameters}

\item{sigsq_beta}{hyperparameters}

\item{a_sigsq}{hyperparameters}

\item{b_sigsq}{hyperparameters}

\item{a_c}{hyperparameters}

\item{b_c}{hyperparameters}

\item{beta_p1_tune}{tuning parameter for adjusted beta proposal}

\item{beta_phi_tune}{tuning parameter for beta proposal}

\item{sigsq_alpha_tune}{tuning parameter for normal proposal}
}
\value{
list of chains for each parameter
}
\description{
fit the model using mcmc
}

