% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locations}
\alias{locations}
\title{Locations Example Data}
\format{A data frame with 11904 observations on the following 3 variables.
\describe{
  \item{x}{
    a numeric vector with the x coordinate of the farm
  }

  \item{y}{
    a numeric vector with the y coordinate of the farm
  }

\item{ppn}{
   The fake id of the farm; equivalent to the source and
    destination fields in the movements dataset
  }
}}
\usage{
data(locations)
}
\description{
Location data included in the package. The data contains a random
sample of actual cattle farm locations in Sweden. This dataset was
obtained from \url{https://www.geodata.se}, the data was last
updated on 2014-05-08 and was downloaded on 2015-11-21. The
download link address is
\url{http://www.jordbruksverket.se/download/18.53b6e8e714255ed1fcc192d/1396620892388/AF.Produktionsplatser.zip}. The
geographic coordinates are in SWEREF99.
}
\keyword{datasets}

