% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mingle.R
\name{make.cond.M.mat}
\alias{make.cond.M.mat}
\title{making transition matrix for network edge probabilities}
\usage{
make.cond.M.mat(p1 = 0.2, phi = 0.95)
}
\arguments{
\item{p1}{the expected density of the network}

\item{phi}{a number between 0 and 1 indicated the stability of the network. 1 results in a fully static network, 0 results in no Markovian dependence at all.}
}
\value{
2x2 matrix filled columnwise with \eqn{p_{0|0}, p_{1|0}, p_{0|1}, p_{1|1}}
}
\description{
making transition matrix for network edge probabilities
}

