% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mingle.R
\name{make.mutil.sim}
\alias{make.mutil.sim}
\title{compute \eqn{\tilde{\boldsymbol{\mu}}(t)} for a given network and set of positions \eqn{\boldsymbol{\mu}(t)}. Intended mostly for internal use in simulation.}
\usage{
make.mutil.sim(W, mu)
}
\arguments{
\item{W}{matrix of dimension \code{c(n.indiv, n.indiv)}}

\item{mu}{matrix of dimension \code{c(n.indiv, 2)}}
}
\value{
matrix of dimension \code{c(n.indiv, 2)}
}
\description{
compute \eqn{\tilde{\boldsymbol{\mu}}(t)} for a given network and set of positions \eqn{\boldsymbol{\mu}(t)}. Intended mostly for internal use in simulation.
}

