\name{mingle-package}
\alias{mingle-package}
\alias{mingle}
\docType{package}
\title{
\packageTitle{mingle}
}
\description{
\packageDescription{mingle}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mingle}
\packageIndices{mingle}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mingle}

Maintainer: \packageMaintainer{mingle}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
