% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mingle.R
\name{sim_mingle}
\alias{sim_mingle}
\title{simulate paths \eqn{\mu} for \code{n.indiv} individuals at \code{T} time points.}
\usage{
sim_mingle(n.indiv = 6, T = 100, params = list(alpha = 0.8, beta = 0.5,
  sigsq = 1, c = 1/3, phi = 0.95, p1 = 0.2), params0 = list(sigsq0 = 30^2),
  seed = 1985)
}
\arguments{
\item{n.indiv}{scalar. number of individuals in network}

\item{T}{number of time points}

\item{params}{list of model parameters used to generate data. List must contain elements \code{alpha, beta, sigsq, c, phi, p1}.}

\item{params0}{list of parameters for initial position \eqn{\mu_0}. Currently only \eqn{\sigma^2_0}.}

\item{seed}{seed to use in \code{set.seed()} before simulating data.}
}
\value{
list of two elements. \code{mu} is an array of dimension \code{c(n.indiv, T, 2)}, and \code{W.sim} is an array of dimension \code{c(n.indiv, n.indiv, T)}.
}
\description{
simulate paths \eqn{\mu} for \code{n.indiv} individuals at \code{T} time points.
}
\examples{
 
simulation.default <- sim_mingle()
}

